Refactor: move static functions to Tools
